+++ src/game/gameobjects/BrickWall.hpp
@@ -70,7 +70,7 @@ public:
 	BrickWall(const BrickWallType & type, const resources::ResourceManager & manager, const glm::vec2 & position, const glm::vec2 & size, const float rotation, const float layer);
 
 	virtual void render() const override;
-	virtual void update(const size_t delta) override;
+	virtual void update(const double delta) override;
 
 private:
 	void renderBrick(const BrickLocation & location) const;
