+++ src/game/gameobjects/Eagle.hpp
@@ -34,6 +34,6 @@ public:
 	Eagle(const resources::ResourceManager & manager, const glm::vec2 & position, const glm::vec2 & size, const float rotation, const float layer);
 
 	virtual void render() const override;
-	virtual void update(const size_t delta) override;
+	virtual void update(const double delta) override;
 };
 }// namespace game
